Separate make Object always return the same data type
This separates CsvObject and EmuseumObject, and has Object
only return what we know will be in both for now.

Its partial, but shows how we can support 'multiple sources of truth'

We may want better strategies where we cache, or allow for a lot more
null, or merge what we think is best.